[NEW] Module rename and inherit payment mode
