+++ l10n_br_account_payment_mode/models/account_move_line.py
@@ -1,0 +1,31 @@
+# -*- encoding: utf-8 -*-
+##############################################################################
+#
+#    Odoo Brazil Account Payment Partner module for Odoo
+#    Copyright (C) 2015 KMEE (http://www.kmee.com.br)
+#    @author Luis Felipe Miléo <mileo@kmee.com.br>
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
+
+from openerp import models, fields, api
+
+
+class AccountMoveLine(models.Model):
+    _inherit = 'account.move.line'
+
+    payment_mode_id = fields.Many2one('payment.mode',
+                                      related='invoice.payment_mode_id',
+                                      string='Payment Mode')
