+++ l10n_br_account_payment_mode/views/payment_view.xml
